requirements.txt
h5py >= 2.2.1,<=2.6.0
numpy >= 1.9.0
biopython >= 1.63
Cython >= 0.25.2
wheel >= 0.29.0
ont_fast5_api == 1.2.0
pysam >= 0.10.0
matplotlib >= 2.0.0
scipy >= 1
torch >= 1

